1. 文章功能优化

+++ ruoyi-modules/ruoyi-system/src/main/java/org/dromara/article/domain/query/ArticleQuery.java
@@ -85,5 +85,5 @@ public class ArticleQuery extends BasePageQuery {
     /**
      * 分类id
      */
-    private Integer sortId;
+    private Long sortId;
 }
